pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ai-prompt-runner"
version = "1.0.0"
description = "CLI tool to send prompts to an AI API and save JSON/Markdown outputs."
requires-python = ">=3.11"
dependencies = [
    "requests>=2.32.5",
    "python-dotenv>=1.2.1",
]

[project.optional-dependencies]
dev = [
    "build==1.3.0",
    "jsonschema[format-nongpl]==4.25.1",
    "pytest==9.0.2",
    "pytest-cov==7.0.0",
    "pytest-sugar==1.1.1",
    "ruff==0.15.1",
]

[project.scripts]
ai-prompt-runner = "ai_prompt_runner.cli:main"

[tool.setuptools.packages.find]
  where = ["src"]
  include = ["ai_prompt_runner*"]
  namespaces = false

[tool.pytest.ini_options]
addopts = "-vv -s"
testpaths = ["tests"]

[tool.ruff]
target-version = "py311"
line-length = 88